Buggy Implementation of history delete
It is now possible to delete an item from the history dashboard. The issue is that the
redirect goes to GET /dashboard instead of GET /dashboard/history. It isn't a big deal,
but I want to take on this challenge. I already have a tentative strategy for it. It is
necessary to modify the delete method in the dashboard controller. When a delete comes
from the dashboard it should POST /dashboard/delete but if the delete comes from history
then it should POST /dashboard/history/delete. This way we can tell both appart. I want
to implement it in such a way that there is only one delete method. I am not sure how to
do that just yet. I am going to have to pass it as a url param somehow.